+++ Mods/SOD.LifeAndLiving/IPluginBindings.cs
@@ -2,109 +2,90 @@
 
 namespace SOD.LifeAndLiving
 {
-    public interface IPluginBindings : IJobBindings, IMurderBindings, ILockPickBindings,
-        IHousingBindings, IMoneyBindings, IShopBindings, IDialogBindings
+    public interface IPluginBindings : IEconomyBindings
     { }
 
-    public interface IDialogBindings
+    public interface IEconomyBindings
     {
-        [Binding(40, "The percentage increase of things such as guest pass cost, things within dialogs.", "LifeAndLiving.Dialog.DialogCostPricePercentage")]
+        [Binding(40, "The percentage increase of things such as guest pass cost, things within dialogs.", "LifeAndLiving.Economy.DialogCostPricePercentage")]
         public int DialogCostPricePercentage { get; set; }
-    }
 
-    public interface IJobBindings
-    {
-        [Binding(90, "The percentage reduction of job resolve payouts.", "LifeAndLiving.Jobs.PayoutReductionJobs")]
+        [Binding(90, "The percentage reduction of job resolve payouts.", "LifeAndLiving.Economy.PayoutReductionJobs")]
         public int PayoutReductionJobs { get; set; }
 
-        [Binding(10, "The minimum amount a side job resolve question should pay (these are all added up to calculate the full side job price).", "LifeAndLiving.Jobs.MinSideJob")]
+        [Binding(10, "The minimum amount a side job resolve question should pay (these are all added up to calculate the full side job price).", "LifeAndLiving.Economy.MinSideJob")]
         public int MinSideJobResolveQuestion { get; set; }
-    }
 
-    public interface IMurderBindings
-    {
-        [Binding(90, "The percentage reduction of murder case resolve payouts.", "LifeAndLiving.Murders.PayoutReductionMurders")]
+        [Binding(90, "The percentage reduction of murder case resolve payouts.", "LifeAndLiving.Economy.PayoutReductionMurders")]
         public int PayoutReductionMurders { get; set; }
-    }
 
-    public interface ILockPickBindings
-    {
-        [Binding(true, "This drastically reduces lockpicks per room", "LifeAndLiving.Lockpicks.ReduceLockPickSpawnPerRoom")]
+        [Binding(true, "This drastically reduces lockpicks per room", "LifeAndLiving.Economy.ReduceLockPickSpawnPerRoom")]
         public bool ReduceLockPickSpawnPerRoom { get; set; }
 
-        [Binding(true, "This drastically reduces lockpicks per address", "LifeAndLiving.Lockpicks.ReduceLockPickSpawnPerAddress")]
+        [Binding(true, "This drastically reduces lockpicks per address", "LifeAndLiving.Economy.ReduceLockPickSpawnPerAddress")]
         public bool ReduceLockPickSpawnPerAddress { get; set; }
 
-        [Binding(true, "This drastically reduces diamonds spawned per address", "LifeAndLiving.Lockpicks.ReduceDiamondSpawnPerAddress")]
+        [Binding(true, "This drastically reduces diamonds spawned per address", "LifeAndLiving.Economy.ReduceDiamondSpawnPerAddress")]
         public bool ReduceDiamondSpawnPerAddress { get; set; }
 
-        [Binding(5, "The amount of lockpicks the buyable kit should give", "LifeAndLiving.Lockpicks.LockPickKitAmount")]
+        [Binding(5, "The amount of lockpicks the buyable kit should give", "LifeAndLiving.Economy.LockPickKitAmount")]
         public int LockPickKitAmount { get; set; }
 
-        [Binding(true, "Should hair pins only spawn in apartment homes?", "LifeAndLiving.Lockpicks.LimitHairPinToHomeOnly")]
+        [Binding(true, "Should hair pins only spawn in apartment homes?", "LifeAndLiving.Economy.LimitHairPinToHomeOnly")]
         public bool LimitHairPinToHomeOnly { get; set; }
 
-        [Binding(true, "Should paper clips only spawn in office buildings?", "LifeAndLiving.Lockpicks.LimitPaperclipToOfficeOnly")]
+        [Binding(true, "Should paper clips only spawn in office buildings?", "LifeAndLiving.Economy.LimitPaperclipToOfficeOnly")]
         public bool LimitPaperclipToOfficeOnly { get; set; }
-    }
 
-    public interface IHousingBindings
-    {
-        [Binding(60, "The percentage increase of apartement cost based on the existing price", "LifeAndLiving.Housing.ApartementCostPercentage")]
+        [Binding(60, "The percentage increase of apartement cost based on the existing price", "LifeAndLiving.Economy.ApartementCostPercentage")]
         public int ApartementCostPercentage { get; set; }
 
-        [Binding(3000, "The minimum $ cost of an apartement, if the percentage increase does not cover this minimum it will be adjusted.", "LifeAndLiving.Housing.MinimumApartementCost")]
+        [Binding(3000, "The minimum $ cost of an apartement, if the percentage increase does not cover this minimum it will be adjusted.", "LifeAndLiving.Economy.MinimumApartementCost")]
         public int MinimumApartementCost { get; set; }
 
-        [Binding(true, "Should diamonds only spawn in apartements.", "LifeAndLiving.Housing.SpawnDiamondsOnlyInApartements")]
+        [Binding(true, "Should diamonds only spawn in apartements.", "LifeAndLiving.Economy.SpawnDiamondsOnlyInApartements")]
         public bool SpawnDiamondsOnlyInApartements { get; set; }
 
-        [Binding(40, "The percentage increase of apartement cost based on the existing price", "LifeAndLiving.Housing.FurniteCostPercentage")]
+        [Binding(40, "The percentage increase of apartement cost based on the existing price", "LifeAndLiving.Economy.FurniteCostPercentage")]
         public int FurniteCostPercentage { get; set; }
 
-        [Binding(250, "The cost of a lower suite in a hotel per day.", "LifeAndLiving.Housing.CostLowerSuiteHotel")]
+        [Binding(250, "The cost of a lower suite in a hotel per day.", "LifeAndLiving.Economy.CostLowerSuiteHotel")]
         public int CostLowerSuiteHotel { get; set; }
 
-        [Binding(500, "The cost of a higher suite in a hotel per day.", "LifeAndLiving.Housing.CostHigherSuiteHotel")]
+        [Binding(500, "The cost of a higher suite in a hotel per day.", "LifeAndLiving.Economy.CostHigherSuiteHotel")]
         public int CostHigherSuiteHotel { get; set; }
-    }
 
-    public interface IMoneyBindings
-    {
-        [Binding(7, "The maximum $ of loose change in a small stack.", "LifeAndLiving.Money.MaxM1Crows")]
+        [Binding(7, "The maximum $ of loose change in a small stack.", "LifeAndLiving.Economy.MaxM1Crows")]
         public int MaxM1Crows { get; set; }
 
-        [Binding(14, "The maximum $ of loose change in a medium stack.", "LifeAndLiving.Money.MaxM2Crows")]
+        [Binding(14, "The maximum $ of loose change in a medium stack.", "LifeAndLiving.Economy.MaxM2Crows")]
         public int MaxM2Crows { get; set; }
 
-        [Binding(18, "The maximum $ of loose change in a big stack.", "LifeAndLiving.Money.MaxM3Crows")]
+        [Binding(18, "The maximum $ of loose change in a big stack.", "LifeAndLiving.Economy.MaxM3Crows")]
         public int MaxM3Crows { get; set; }
 
-        [Binding(25, "The maximum $ of loose change in a huge stack.", "LifeAndLiving.Money.MaxM4Crows")]
+        [Binding(25, "The maximum $ of loose change in a huge stack.", "LifeAndLiving.Economy.MaxM4Crows")]
         public int MaxM4Crows { get; set; }
 
-        [Binding(true, "Should loose change only spawn in apartements.", "LifeAndLiving.Money.LimitLooseMoneyToApartementsOnly")]
+        [Binding(true, "Should loose change only spawn in apartements.", "LifeAndLiving.Economy.LimitLooseMoneyToApartementsOnly")]
         public bool LimitLooseMoneyToApartementsOnly { get; set; }
-    }
 
-    public interface IShopBindings
-    {
-        [Binding(60, "The percentage the value should increase of items.", "LifeAndLiving.Shop.PercentageValueIncrease")]
+        [Binding(60, "The percentage the value should increase of items.", "LifeAndLiving.Economy.PercentageValueIncrease")]
         public int PercentageValueIncrease { get; set; }
 
-        [Binding(32, "The minimum value of an item.", "LifeAndLiving.Shop.MinItemValue")]
+        [Binding(32, "The minimum value of an item.", "LifeAndLiving.Economy.MinItemValue")]
         public int MinItemValue { get; set; }
 
-        [Binding(25, "The percentage of item value that is taken for selling items to general stores (default game is 50%)", "LifeAndLiving.Shop.PercentageSalePriceGeneral")]
+        [Binding(25, "The percentage of item value that is taken for selling items to general stores (default game is 50%)", "LifeAndLiving.Economy.PercentageSalePriceGeneral")]
         public int PercentageSalePriceGeneral { get; set; }
 
-        [Binding(50, "The percentage of item value that is taken for selling items to blackmarket (default game is 80%)", "LifeAndLiving.Shop.PercentageSalePriceBlackMarket")]
+        [Binding(50, "The percentage of item value that is taken for selling items to blackmarket (default game is 80%)", "LifeAndLiving.Economy.PercentageSalePriceBlackMarket")]
         public int PercentageSalePriceBlackMarket { get; set; }
 
-        [Binding(400, "The minimum value of a diamond.", "LifeAndLiving.Shop.MinDiamondValue")]
+        [Binding(400, "The minimum value of a diamond.", "LifeAndLiving.Economy.MinDiamondValue")]
         public int MinDiamondValue { get; set; }
 
-        [Binding(1000, "The maximum value of a diamond.", "LifeAndLiving.Shop.MaxDiamondValue")]
+        [Binding(1000, "The maximum value of a diamond.", "LifeAndLiving.Economy.MaxDiamondValue")]
         public int MaxDiamondValue { get; set; }
     }
 }