+++ Mods/SOD.LifeAndLiving/ModFolderContent/CHANGELOG.md
@@ -1,3 +1,14 @@
 # CHANGELOG
 **1.0.0**
-- Initial release
+- Reduce payout of side jobs
+- Reduce payout of murder cases
+- Reduce lockpick amount received from buying a lockpick kit
+- Reduce spawn rate of lockpicks
+- Reduce the spawn rate of diamonds
+- Reduce the value of loose change
+- Reduce the spawn rate of loose change
+- Increase cost of apartment
+- Increase cost of hotel suites
+- Adjust item value and store prices
+- Adjust cost price of furniture
+- Adjust cost price of dialog (guest pass, codes, etc..)